+++ src/test/java/com/pahod/repository/storage/CommonInMemoryStorageTest.java
@@ -1,9 +1,7 @@
 package com.pahod.repository.storage;
 
 import com.pahod.model.Event;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Spy;
 
 import java.util.Map;
 
@@ -16,14 +14,14 @@ class CommonInMemoryStorageTest {
         //given
         CommonInMemoryStorage storage = new CommonInMemoryStorage();
 
-        int lastId = 15;
+        long lastId = 15;
         Class<Event> className = Event.class;
 
-        Map<Class<?>, Integer> lastIdMap = storage.getLastIdForClass();
+        Map<Class<?>, Long> lastIdMap = storage.getLastIdForClass();
         lastIdMap.put(className, lastId);
 
         //when
-        int generatedId = storage.generateNextIdForClass(className);
+        long generatedId = storage.generateNextIdForClass(className);
 
         //then
         assertEquals(generatedId, lastId + 1);
